double -> float

+++ inc/libgxns.h
@@ -6,7 +6,7 @@
 /*   By: njaber <neyl.jaber@gmail.com>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/01/25 23:56:43 by njaber            #+#    #+#             */
-/*   Updated: 2018/05/20 01:44:27 by njaber           ###   ########.fr       */
+/*   Updated: 2018/05/29 03:29:07 by njaber           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,14 +55,14 @@ typedef void	*t_ocl;
 # endif
 
 typedef struct	s_vec2 {
-	double	x;
-	double	y;
+	float	x;
+	float	y;
 }				t_vec2;
 
 typedef struct	s_vec3 {
-	double	x;
-	double	y;
-	double	z;
+	float	x;
+	float	y;
+	float	z;
 }				t_vec3;
 
 typedef float	t_mat4[16];
@@ -88,7 +88,7 @@ typedef struct	s_win {
 	t_img				img;
 	t_ivec				size;
 	int					frame;
-	double				fps;
+	float				fps;
 	unsigned long		frames[30];
 }				t_win;
 
@@ -101,7 +101,7 @@ void			scale(t_mat4 m, t_vec3 v);
 void			rotate(t_mat4 m, t_vec3 v);
 void			multiply(t_mat4 m1, t_mat4 m2, int stock_in_m2);
 
-double			sqr(double n);
+float			sqr(float n);
 
 t_vec2			c_mult(t_vec2 z1, t_vec2 z2);
 t_vec2			c_add(t_vec2 z1, t_vec2 z2);
@@ -109,14 +109,14 @@ t_vec2			c_pow(t_vec2 z1, int pow);
 
 t_vec3			vec_sub(t_vec3 v1, t_vec3 v2);
 t_vec3			vec_add(t_vec3 v1, t_vec3 v2);
-t_vec3			vec_mult(t_vec3 v1, double d);
-double			length(t_vec3 v1);
-double			scalar(t_vec3 v1, t_vec3 v2);
+t_vec3			vec_mult(t_vec3 v1, float d);
+float			length(t_vec3 v1);
+float			scalar(t_vec3 v1, t_vec3 v2);
 
 int				init_new_win(void *mlx, t_win *win, t_ivec size, char *title);
 void			paint_window(t_win *win, t_kernel *opencl_kernel);
 
-void			display_data_float(t_win *win, char *name, double data, int y);
+void			display_data_float(t_win *win, char *name, float data, int y);
 void			display_data_vec2(t_win *win, char *name, t_vec2 data, int y);
 void			display_data_vec3(t_win *win, char *name, t_vec3 data, int y);
 void			display_data_str(t_win *win, char *name, char *data, int y);
@@ -126,7 +126,7 @@ void			img_px(t_img *img, unsigned int color, t_ivec pixel);
 void			clear_img(t_img *img);
 
 void			draw_line(t_img *img, unsigned int *color, t_vec2 *verticies);
-unsigned int	color_gradiant(unsigned int *colors, double gradiant);
+unsigned int	color_gradiant(unsigned int *colors, float gradiant);
 
 int				get_key_digit(int key_code);
