+++ libgxns/matrix2.c
@@ -15,7 +15,7 @@
 t_vec3		apply_mat_vec3(t_vec3 v, t_mat4 m)
 {
 	t_vec3	ret;
-	double	w;
+	float	w;
 
 	w = m[12] * v.x + m[13] * v.y + m[14] * v.z + m[15];
 	ret.x = (m[0] * v.x + m[1] * v.y + m[2] * v.z + m[3]) / w;
